requirements.txt
#wheel
wheel==0.37.1
#devops
black==22.3.0
click==8.1.3 
pytest==7.1.3
pytest-cov==4.0.0
pylint==2.15.3
boto3==1.24.87
#web
fastapi==0.85.0
uvicorn==0.18.3
#datascience
jupyter
pandas
numpy
pyarrow
ipynb
black_nbconvert
ruff
nbval
plotly
nbqa
pytest-playwright
flask
flask_mail
sqlalchemy
flask_sqlalchemy
finnhub-python

#Connect to DB
databricks
databricks-sql-connector